+++ src/DoctrineMongoODMModule/Module.php
@@ -83,7 +83,6 @@ class Module
         return [
             'invokables' => [
                 Logging\DebugStack::class  => Logging\DebugStack::class,
-                Logging\LoggerChain::class => Logging\LoggerChain::class,
                 Logging\EchoLogger::class  => Logging\EchoLogger::class,
             ],
             'aliases' => [DocumentManager::class => 'doctrine.documentmanager.odm_default'],