+++ CHANGELOG.md
@@ -24,6 +24,8 @@
 - `DoctrineMongoODMModule\Paginator\Adapter\DoctrinePaginator` was removed because `Doctrine\ODM\MongoDB\Cursor` doesn't exist anymore and there is no simple workaround.
 - The CLI commands `odm:generate:documents` and `odm:generate:repositories` were removed because they no longer exist.
 - The Configuration options `retryConnect` and `retryQuery` were removed.
+- The `DoctrineMongoODMModule\Logging\Logger` interface was removed. Existing implementations directly implement `\Doctrine\ODM\MongoDB\APM\CommandLoggerInterface` now.
+- The `DoctrineMongoODMModule\Logging\LoggerChain` was removed. Loggers can now be subscribed with the MongoDB driver directly.
 
 ### Fixed